feat: add map intel panel, ancient Elo badges, and F3 overlay toggle
- Map intel panel (F3 toggle): shows per-character per-act damage stats,
  per-encounter elite/boss breakdown, win rates, elite count correlation
- Ancient choice Elo badges on NEventOptionButton via Harmony patch
- F3 global toggle via NGame._Input Harmony patch (replaces broken Node approach)
- Fix card overlay positioning (CenterBottom) and flickering (MouseFilter.Ignore)
- Export per-encounter damage, per-act win rates, elite correlation to overlay JSON

+++ src/Sts2Analytics.Core/Elo/AncientRatingEngine.cs
@@ -51,8 +51,8 @@ public class AncientRatingEngine
             foreach (var floorGroup in choicesByFloor)
             {
                 var floorChoices = floorGroup.ToList();
-                var picked = floorChoices.Where(c => c.WasChosen != 0).ToList();
-                var skipped = floorChoices.Where(c => c.WasChosen == 0).ToList();
+                var picked = floorChoices.Where(c => (long)c.WasChosen != 0).ToList();
+                var skipped = floorChoices.Where(c => (long)c.WasChosen == 0).ToList();
 
                 if (picked.Count == 0 || skipped.Count == 0) continue;
